helpers added to core file
